Code refactoring for all controllers
- Removing unnecessary methodes from controllers
- Profesion and Location controllers do not belong to users routes anymore
- Document controller now belongs to users route and it is stored in users folder structure
- Changing profession and location links to remove 'users.' prefix
- Protecting users routes with middleware added to route groups
- Constructors are removed
