+++ app/Http/Controllers/Admins/LocationController.php
@@ -51,17 +51,6 @@ class LocationController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
@@ -88,17 +77,6 @@ class LocationController extends Controller
 
         return redirect()->route('admins.locations.index')
                          ->withStatus("Location '{$location->name}' has been updated successfully.");
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy($id)
-    {
-        //
     }
 
     public function candidates(Location $location)
