--- app/Http/Controllers/Users/LocationController.php
@@ -1,100 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Users;
-
-use App\Models\Location;
-use Illuminate\Http\Request;
-use App\Facades\CounterFacade;
-use App\Http\Controllers\Controller;
-
-class LocationController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        return view ('users.locations.index', [
-            'locations' => Location::all()
-        ]);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Location  $location
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Location $location)
-    {
-        $professions = $location->professions()
-                                ->withoutExpiredProfessions() // exclude expired professions
-                                ->with('locations')
-                                ->paginate(10);
-
-        $counter = CounterFacade::increment("location-{$location->id}", ["location"]);
-
-        return view('users.locations.show', [
-            'location' => $location,
-            'professions' => $professions,
-            'counter' => $counter,
-        ]);
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Location  $location
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Location $location)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Location  $location
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Location $location)
-    {
-        //
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Location  $location
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Location $location)
-    {
-        //
-    }
-}
