+++ resources/views/components/profession-results.blade.php
@@ -9,7 +9,7 @@
       @elseif (Route::is('admins.candidates.professions.results'))
         <a href="{{ route('admins.candidates.show', ['candidate' => $value->candidate->id]) }}" class="text-decoration-none">{{ $value->candidate->user->email }}</a>
       @else
-        <a href="{{ route('users.professions.show', ['profession' => $value->profession->id]) }}" class="text-decoration-none">{{ $value->profession->title }}</a>
+        <a href="{{ route('professions.show', ['profession' => $value->profession->id]) }}" class="text-decoration-none">{{ $value->profession->title }}</a>
       @endif
       result, attempted {{ $value->updated_at->diffForHumans() }}.
       @if ($value->trashed())
