+++ resources/views/layouts/admin.blade.php
@@ -66,7 +66,7 @@
       <div class="bg-light border-right" id="sidebar-wrapper">
         <div class="sidebar-heading">Admin panel</div>
         <div class="list-group list-group-flush">
-            <a href="{{ route('users.professions.index') }}" class="list-group-item list-group-item-action bg-light">Home</a>
+            <a href="{{ route('professions.index') }}" class="list-group-item list-group-item-action bg-light">Home</a>
             <div class="dropdown">
               <a href="#" class="dropdown-toggle list-group-item list-group-item-action bg-light" id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">Admin users</a>
               <ul class="dropdown-menu" aria-labelledby="dropdownMenuButton1">
