+++ resources/views/locations/show.blade.php
