+++ resources/views/professions/index.blade.php
