+++ resources/views/professions/show.blade.php
