+++ resources/views/users/candidates/professions/show.blade.php
@@ -4,7 +4,7 @@
     
 @section('content')
   <h1>
-    <a href="{{ route('users.professions.show', ['profession' => $profession->id]) }}" class="text-decoration-none">{{ $profession->title }}</a>
+    <a href="{{ route('professions.show', ['profession' => $profession->id]) }}" class="text-decoration-none">{{ $profession->title }}</a>
   </h1>
   <div class="mb-5">
     <p>
